requirements.txt
django 
djangorestframework 
django-cors-headers
django-rest_framework